OLED: Add OLED_COLUMN_OFFSET support for SSD1306
There are some 64×32 and 64×48 OLED displays with SSD1306-like
controllers that need the `OLED_COLUMN_OFFSET` feature (the 64 SEG
outputs used for the OLED matrix are in the middle of the 128 SEG
outputs supported by the controller).  Adding `OLED_COLUMN_OFFSET` to
the column address makes these displays usable with some restrictions
(e.g., scrolling does not work properly, because the uninitialized
portions of display memory are scrolled into view, and proper rotation
for display sizes like 64×48 needs further fixes).