Correct key events time resolution to 1ms
The bit wise OR that was used before would always set the least significant
bit to one.  So the timestamp could never be a even value. Therefore the
check if the current key event is really a product of a key press/release
is now implemented as a simple flag on the keyevent_t type.